[AR-3694] Get specs to pass on build server
There is a spec that will not pass on the build server. It involves
having a directory that is not owned nor accessible by the current user.
This test scenario cannot be setup as is on the build server since the
build server runs as root (or jenkins which has all the access privledges
of root). Therefore, this spec will be skipped on the build server but
will be ran by local developers.

Concession: The specs in funzip are lacking in error scenarios
anyway. This change leaves the whole thing in no less of a bad state as
it already was - no running on the build server at all but only on dev
machines.